Moved RedisForwardingSink to common module
