+++ common/src/main/java/com/vexsoftware/votifier/support/forwarding/redis/RedisForwardingSink.java
@@ -1,12 +1,10 @@
-package com.vexsoftware.votifier.forwarding;
+package com.vexsoftware.votifier.support.forwarding.redis;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.vexsoftware.votifier.model.Vote;
 import com.vexsoftware.votifier.support.forwarding.ForwardedVoteListener;
 import com.vexsoftware.votifier.support.forwarding.ForwardingVoteSink;
-import com.vexsoftware.votifier.support.forwarding.redis.RedisCredentials;
-import com.vexsoftware.votifier.support.forwarding.redis.RedisPoolConfiguration;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
 import redis.clients.jedis.JedisPoolConfig;
