+++ sponge/src/main/java/com/vexsoftware/votifier/sponge/NuVotifier.java
@@ -14,12 +14,10 @@ import com.vexsoftware.votifier.sponge.cmd.NVReloadCmd;
 import com.vexsoftware.votifier.sponge.cmd.TestVoteCmd;
 import com.vexsoftware.votifier.sponge.config.ConfigLoader;
 import com.vexsoftware.votifier.sponge.event.VotifierEvent;
-import com.vexsoftware.votifier.sponge.forwarding.RedisForwardingSink;
 import com.vexsoftware.votifier.sponge.forwarding.SpongePluginMessagingForwardingSink;
 import com.vexsoftware.votifier.support.forwarding.ForwardedVoteListener;
 import com.vexsoftware.votifier.support.forwarding.ForwardingVoteSink;
-import com.vexsoftware.votifier.support.forwarding.redis.RedisCredentials;
-import com.vexsoftware.votifier.support.forwarding.redis.RedisPoolConfiguration;
+import com.vexsoftware.votifier.support.forwarding.redis.*;
 import com.vexsoftware.votifier.util.KeyCreator;
 import org.slf4j.Logger;
 import org.spongepowered.api.Sponge;
